Add mask email helper class
